requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.2.4
django-bootstrap4==23.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.6.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.3.1
redis==5.0.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==3.0.1
zope.interface==6.0
